+++ plugins/wazuh-analysis/public/groups/cloud-security/apps/aws/aws-app.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { AppMountParameters } from 'opensearch-dashboards/public';
-import { AWS_TITLE } from '../../constants';
+import { AWS_TITLE } from './constants';
 
 interface AwsAppProps {
   params: AppMountParameters;
