+++ plugins/wazuh-analysis/public/groups/cloud-security/apps/google-cloud/application.tsx
@@ -3,9 +3,10 @@ import { AppMountParameters } from 'opensearch-dashboards/public';
 import ReactDOM from 'react-dom';
 import { Layout } from '../../../layout';
 import { createSideNavItems } from '../../../side-nav';
-import { CLOUD_SECURITY_TITLE, GOOGLE_CLOUD_ID } from '../../constants';
+import { CLOUD_SECURITY_TITLE } from '../../constants';
 import { CloudSecurityNavGroup } from '../..';
 import { GoogleCloudApp } from './google-cloud-app';
+import { GOOGLE_CLOUD_ID } from './constants';
 
 export const renderApp = async (params: AppMountParameters) => {
   const items = createSideNavItems({
